added the wheelchair functionality
With the use of 2D arrays to store the location of the buttons. for the wheelchair user you can take out an adjacent seat out and then that seat becomes unavailable.Also seats that are at the end of the edges are also invalid to take seats out from and will be given an error.